Refactor: Move signature visitors to `signature` module
