+++ src/asm/mod.rs
@@ -3,3 +3,4 @@ pub mod generate;
 pub mod node;
 #[cfg(feature = "parse")]
 pub mod parse;
+pub mod visitor;
